+++ gateway/src/backend/start.rs
@@ -1,4 +1,4 @@
-use crate::backend_commands::BackendCommands;
+use crate::backend::commands::BackendCommands;
 use crate::kafka_io;
 use crate::redis_io;
 use redis_io::RedisPool;
