--- gateway/src/backend_events.rs
@@ -1,1 +1,0 @@
-
